split CheckVersionDB to a single module
CheckVersionDB is shared by two files, we want Avocado to copy
this file to VM in artifact tests.

(cherry picked from commit af515347ab1ae5e321b449073bdbd06f86603bec)
